requirements.txt
fastapi==0.118.0
uvicorn==0.37.0
sqlalchemy==2.0.43
pymysql==1.1.2
python-jose==3.5.0
passlib[bcrypt]==1.7.4
python-dotenv==1.1.1
pydantic[email]==2.11.10
alembic==1.16.5
apscheduler==3.11.0
pytest==8.4.2
httpx==0.28.1
bcrypt==4.0.1
passlib==1.7.4
mysql-connector-python
aiomysql
pytest-asyncio
cryptography
pytz
pyinstrument